requirements.txt
aiohttp>=3.9.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
pyupbit>=0.2.33
websockets>=12.0
tenacity>=8.2.3
loguru>=0.7.2
